Changed PC advancement so it occurs immediately upon a fetch within INstructionManager.
I think for sending data to the UI, you can just create a UIManager object inside the OperationManager. Whenever a register value is updated, we call UIManager.UpdateRegister(rd.GetName())

There will be a map in UIManager that contains UI text objects or something and it will have a key value of string.